requirements.txt
google-genai
python-dotenv
pyaudio
